video_processor/requirements.txt
requests
torch
einops
timm
boto3
opencv-python-headless
Pillow
numpy
python-dotenv
kafka-python
pytest